checkpoint: Rename `affils` app to `affiliations`
