+++ src/affiliations/urls.py
@@ -1,10 +1,10 @@
-"""URLs for the the affiliations service."""
+"""URLs for the affiliations service."""
 
 # Third-party dependencies:
 from django.urls import path
 
 # In-house code:
-from affils import views
+from affiliations import views
 
 urlpatterns = [
     path("", views.affiliations_list),
