+++ src/main/urls.py
@@ -9,6 +9,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path("", include("affils.urls")),
+    path("", include("affiliations.urls")),
     path("admin/", admin.site.urls),
 ]
